Refactor CloudWatch
